Removed some useless dependencies and renamed files for clarity
