+++ package.json
@@ -32,13 +32,10 @@
     "rehype-autolink-headings": "^7.1.0",
     "rehype-pretty-code": "^0.14.1",
     "rehype-slug": "^6.0.0",
-    "remark-code-blocks": "^2.0.1",
     "remark-gfm": "^4.0.1",
     "remark-parse": "^11.0.0",
-    "shiki": "^3.8.1",
     "tailwind-merge": "^3.0.2",
     "tailwindcss-animate": "^1.0.7",
-    "tw-to-css": "^0.0.12",
     "unified": "^11.0.5",
     "unist-util-visit": "^5.0.0"
   },
@@ -50,7 +47,6 @@
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
-    "@types/react-syntax-highlighter": "^15.5.13",
     "@types/unist": "^3.0.3",
     "eslint": "^9",
     "eslint-config-next": "15.1.4",
